Separated setting up the network and running it into two different files
