+++ .gitignore
@@ -1,65 +1,2 @@
-build/cmake_install.cmake
-build/CMakeCache.txt
-build/compile_commands.json
-build/Makefile
-build/ml_exec
-build/.cmake/api/v1/query/client-vscode/query.json
-build/.cmake/api/v1/reply/cache-v2-67d44788866d69f0528e.json
-build/.cmake/api/v1/reply/index-2025-07-28T07-08-28-0404.json
-build/.cmake/api/v1/reply/cmakeFiles-v1-03e048ea42532307d425.json
-build/.cmake/api/v1/reply/codemodel-v2-f6e0443095811a80723d.json
-build/.cmake/api/v1/reply/directory-.-Debug-f5ebdc15457944623624.json
-build/.cmake/api/v1/reply/index-2025-07-22T11-25-56-0668.json
-build/.cmake/api/v1/reply/target-ml_exec-Debug-667c5397f82c9dbc6acc.json
-build/.cmake/api/v1/reply/toolchains-v1-de13eb4a293a7f94668f.json
-build/.cmake/api/v1/reply/index-2025-07-29T21-04-20-0609.json
-build/CMakeFiles/cmake.check_cache
-build/CMakeFiles/CMakeConfigureLog.yaml
-build/CMakeFiles/CMakeDirectoryInformation.cmake
-build/CMakeFiles/Makefile.cmake
-build/CMakeFiles/Makefile2
-build/CMakeFiles/progress.marks
-build/CMakeFiles/TargetDirectories.txt
-build/CMakeFiles/3.28.3/CMakeCCompiler.cmake
-build/CMakeFiles/3.28.3/CMakeCXXCompiler.cmake
-build/CMakeFiles/3.28.3/CMakeDetermineCompilerABI_C.bin
-build/CMakeFiles/3.28.3/CMakeDetermineCompilerABI_CXX.bin
-build/CMakeFiles/3.28.3/CMakeSystem.cmake
-build/CMakeFiles/3.28.3/CompilerIdC/a.out
-build/CMakeFiles/3.28.3/CompilerIdC/CMakeCCompilerId.c
-build/CMakeFiles/3.28.3/CompilerIdCXX/a.out
-build/CMakeFiles/3.28.3/CompilerIdCXX/CMakeCXXCompilerId.cpp
-build/CMakeFiles/launch.dir/build.make
-build/CMakeFiles/launch.dir/cmake_clean.cmake
-build/CMakeFiles/launch.dir/compiler_depend.make
-build/CMakeFiles/launch.dir/compiler_depend.ts
-build/CMakeFiles/launch.dir/depend.make
-build/CMakeFiles/launch.dir/DependInfo.cmake
-build/CMakeFiles/launch.dir/flags.make
-build/CMakeFiles/launch.dir/imgProc.c.o
-build/CMakeFiles/launch.dir/imgProc.c.o.d
-build/CMakeFiles/launch.dir/link.txt
-build/CMakeFiles/launch.dir/main.c.o
-build/CMakeFiles/launch.dir/main.c.o.d
-build/CMakeFiles/launch.dir/neural.c.o
-build/CMakeFiles/launch.dir/neural.c.o.d
-build/CMakeFiles/launch.dir/progress.make
-build/CMakeFiles/ml_exec.dir/build.make
-build/CMakeFiles/ml_exec.dir/cmake_clean.cmake
-build/CMakeFiles/ml_exec.dir/compiler_depend.internal
-build/CMakeFiles/ml_exec.dir/compiler_depend.make
-build/CMakeFiles/ml_exec.dir/compiler_depend.ts
-build/CMakeFiles/ml_exec.dir/depend.make
-build/CMakeFiles/ml_exec.dir/DependInfo.cmake
-build/CMakeFiles/ml_exec.dir/flags.make
-build/CMakeFiles/ml_exec.dir/link.txt
-build/CMakeFiles/ml_exec.dir/progress.make
-build/CMakeFiles/ml_exec.dir/src/imgProc.c.o
-build/CMakeFiles/ml_exec.dir/src/imgProc.c.o.d
-build/CMakeFiles/ml_exec.dir/src/main.c.o
-build/CMakeFiles/ml_exec.dir/src/main.c.o.d
-build/CMakeFiles/ml_exec.dir/src/mathFun.c.o
-build/CMakeFiles/ml_exec.dir/src/mathFun.c.o.d
-build/CMakeFiles/ml_exec.dir/src/neural.c.o
-build/CMakeFiles/ml_exec.dir/src/neural.c.o.d
+build/
 *.csv
