+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "files.associations": {
         ".fantomasignore": "ignore",
         "imgproc.h": "c",
-        "math.h": "c"
+        "math.h": "c",
+        "neural.h": "c"
     }
 }
